fix(saas): update record input props
